Refactor services
Split the service module into encryptor_service and aws_service, so that
it doesn't become a dumping ground.  This is in preparation for adding
another module that talks to the Bracket service.

Test plan: Ran unit tests, encrypted an AMI.
