+++ brkt_cli/aws_service.py
@@ -17,16 +17,9 @@ import re
 import boto
 import boto.sts
 import logging
-import json
-import urllib2
 
 from boto.exception import EC2ResponseError
 import time
-
-ENCRYPT_SUCCESSFUL = 'finished'
-ENCRYPT_FAILED = 'failed'
-ENCRYPTOR_STATUS_PORT = 8000
-FAILURE_CODE_UNSUPPORTED_GUEST = 'unsupported_guest'
 
 PLATFORM_WINDOWS = 'windows'
 
@@ -384,47 +377,6 @@ class AWSService(BaseAWSService):
 
     def get_console_output(self, instance_id):
         return self.conn.get_console_output(instance_id)
-
-
-class BaseEncryptorService(object):
-    __metaclass__ = abc.ABCMeta
-
-    def __init__(self, hostname, port=ENCRYPTOR_STATUS_PORT):
-        self.hostname = hostname
-        self.port = port
-
-    @abc.abstractmethod
-    def is_encryptor_up(self):
-        pass
-
-    @abc.abstractmethod
-    def get_status(self):
-        pass
-
-
-class EncryptorService(BaseEncryptorService):
-
-    def is_encryptor_up(self):
-        try:
-            self.get_status()
-            return True
-        except Exception as e:
-            log.debug("Couldn't get encryptor status: %s", e)
-            return False
-
-    def get_status(self, timeout_secs=2):
-        url = 'http://%s:%d/encryption_status' % (self.hostname, self.port)
-        r = urllib2.urlopen(url, timeout=timeout_secs)
-        data = r.read()
-        info = json.loads(data)
-        ratio = 0
-        info['percent_complete'] = 0
-        if info['state'] == ENCRYPT_SUCCESSFUL:
-            info['percent_complete'] = 100
-        elif info['bytes_total'] > 0:
-            ratio = float(info['bytes_written']) / info['bytes_total']
-            info['percent_complete'] = int(100 * ratio)
-        return info
 
 
 class ImageNameError(Exception):
